Moved create_test_data to xarray.testing
