+++ xarray/tests/test_combine.py
@@ -13,11 +13,11 @@ from xarray.core.combine import (
     _check_shape_tile_ids, _combine_nd, _infer_concat_order_from_positions,
     _infer_tile_ids_from_nested_list, _new_tile_id)
 from xarray.core.pycompat import OrderedDict, iteritems
+from xarray.testing import create_test_data
 
 from . import (
     InaccessibleArray, assert_array_equal, assert_combined_tile_ids_equal,
     assert_equal, assert_identical, raises_regex, requires_dask)
-from .test_dataset import create_test_data
 
 
 class TestConcatDataset(object):
