+++ xarray/tests/test_dataset.py
@@ -20,6 +20,7 @@ from xarray.core import dtypes, indexing, npcompat, utils
 from xarray.core.common import full_like
 from xarray.core.pycompat import (
     OrderedDict, integer_types, iteritems, unicode_type)
+from xarray.testing import create_test_data
 
 from . import (
     InaccessibleArray, UnexpectedDataAccess, assert_allclose,
@@ -31,27 +32,6 @@ try:
     import dask.array as da
 except ImportError:
     pass
-
-
-def create_test_data(seed=None):
-    rs = np.random.RandomState(seed)
-    _vars = {'var1': ['dim1', 'dim2'],
-             'var2': ['dim1', 'dim2'],
-             'var3': ['dim3', 'dim1']}
-    _dims = {'dim1': 8, 'dim2': 9, 'dim3': 10}
-
-    obj = Dataset()
-    obj['time'] = ('time', pd.date_range('2000-01-01', periods=20))
-    obj['dim2'] = ('dim2', 0.5 * np.arange(_dims['dim2']))
-    obj['dim3'] = ('dim3', list('abcdefghij'))
-    for v, dims in sorted(_vars.items()):
-        data = rs.normal(size=tuple(_dims[d] for d in dims))
-        obj[v] = (dims, data, {'foo': 'variable'})
-    obj.coords['numbers'] = ('dim3', np.array([0, 1, 2, 0, 0, 1, 1, 2, 2, 3],
-                                              dtype='int64'))
-    obj.encoding = {'foo': 'bar'}
-    assert all(obj.data.flags.writeable for obj in obj.variables.values())
-    return obj
 
 
 def create_test_multiindex():
