+++ xarray/tests/test_distributed.py
@@ -18,7 +18,7 @@ from xarray.backends.locks import HDF5_LOCK, CombinedLock
 from xarray.tests.test_backends import (ON_WINDOWS, create_tmp_file,
                                         create_tmp_geotiff,
                                         open_example_dataset)
-from xarray.tests.test_dataset import create_test_data
+from xarray.testing import create_test_data
 
 from . import (
     assert_allclose, has_h5netcdf, has_netCDF4, requires_rasterio, has_scipy,
