+++ xarray/tests/test_interp.py
@@ -7,10 +7,10 @@ import pytest
 import xarray as xr
 from xarray.tests import (
     assert_allclose, assert_equal, requires_cftime, requires_scipy)
+from xarray.testing import create_test_data
 
 from . import has_dask, has_scipy
 from ..coding.cftimeindex import _parse_array_of_cftime_strings
-from .test_dataset import create_test_data
 
 try:
     import scipy
